+++ youtube_downloader/__init__.py
@@ -1,3 +1,3 @@
 """YouTube Downloader V2 package."""
 
-__version__ = "2.0.0"
+__version__ = "2.0.1"